fix(eevee): store outline info as integer image

+++ source/blender/draw/engines/eevee/shaders/eevee_outline_freestyle_frag.glsl
@@ -15,7 +15,7 @@ float4 outline_source_color_fetch(int2 texel)
   return texelFetch(outline_color_tx, texel, 0);
 }
 
-float4 outline_source_info_fetch(int2 texel)
+uint4 outline_source_info_fetch(int2 texel)
 {
   return texelFetch(outline_info_tx, texel, 0);
 }
@@ -40,7 +40,7 @@ void main()
   }
 
   const float4 outline_color = outline_source_color_fetch(texel);
-  const float4 outline_info = outline_source_info_fetch(texel);
+  const uint4 outline_info = outline_source_info_fetch(texel);
   const float line_width = outline_width_unpack(outline_info.r);
   if (line_width <= 0.0f || outline_color.a <= 0.0f ||
       !outline_freestyle_edge_unpack(outline_info.b))
